Move slurm to utils
